refactor: Calendar -> DisplayCalendar 이름 변경 및 constant, 함수 분리
