+++ src/components/Calendar/displayCalendar.css
@@ -1,0 +1,10 @@
+@import '../../styles/calendar.css';
+
+.react-calendar {
+  box-shadow: 1px 1px 6px 1px rgba(0, 0, 0, 0.25);
+}
+
+/* month-view 스타일 */
+.react-calendar__month-view__days {
+  pointer-events: none;
+}
