+++ src/constants/constants.ts
@@ -1,0 +1,1 @@
+export const WEEK_DAYS = ['S', 'M', 'T', 'W', 'T', 'F', 'S'];
