+++ src/styles/calendar.css
@@ -2,22 +2,37 @@
   background: #fff;
   color: black;
   width: 231px;
+  max-width: 100%;
   font-size: 0.75rem;
-  background: none;
+  font-weight: bold;
   line-height: normal;
   border: none;
   border-radius: 10px;
-  box-shadow: 1px 1px 6px 1px rgba(0, 0, 0, 0.25);
   box-sizing: border-box;
 }
 
-.react-calendar__month-view__days {
-  pointer-events: none;
+/* tile 스타일 */
+.react-calendar__year-view .react-calendar__tile,
+.react-calendar__decade-view .react-calendar__tile,
+.react-calendar__century-view .react-calendar__tile {
+  padding: 1em 0.5em;
 }
 
+.react-calendar__tile {
+  padding: 5px;
+}
+
+/* navigation */
 .react-calendar__navigation {
+  display: flex;
   height: 35px;
-  margin-bottom: 0px;
+  margin-bottom: 1rem;
+  font-size: 1rem;
+}
+
+.react-calendar__navigation button {
+  min-width: 44px;
+  background: none;
 }
 
 .react-calendar__navigation__prev2-button,
@@ -25,20 +40,26 @@
   display: none;
 }
 
-.react-calendar__tile--now {
-  background: none;
-}
-
-.react-calendar__tile {
-  padding: 5px 3px;
+/* month-view 스타일 */
+.react-calendar__month-view {
+  padding: 0 1rem;
 }
 
 .react-calendar__month-view__weekdays {
-  font-size: 0.75rem;
+  text-align: center;
+  text-transform: uppercase;
+  font-weight: bold;
+  font-size: 0.75em;
+  margin-bottom: 0.44rem;
+}
+
+.react-calendar__month-view__weekdays__weekday {
+  padding: 0.5em;
 }
 
 .react-calendar__month-view__weekdays__weekday abbr {
   text-decoration: none;
+  font-size: 0.75rem;
 }
 
 .react-calendar__month-view__weekdays__weekday:first-of-type {
@@ -50,9 +71,11 @@
 }
 
 .react-calendar__month-view__days {
-  margin-bottom: 5px;
+  margin-bottom: 0.56rem;
+  gap: 0.19rem 0;
 }
 
+/* 주말(토/일) 색상 스타일 */
 .react-calendar__month-view__days__day--weekend {
   color: var(--main-color);
 }
@@ -60,18 +83,23 @@
 .react-calendar__month-view__days
   button:nth-of-type(7n):not(
     .react-calendar__month-view__days__day--neighboringMonth
+  ):not(.selected-continuous):not(.selected-continuous-first):not(
+    .selected-continuous-last
   ) {
   color: blue;
 }
 
-.selected::before {
+/* 선택된 날짜 스타일 */
+.selected:not(
+    .react-calendar__month-view__days__day--neighboringMonth
+  )::before {
   content: '';
   position: absolute;
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
-  width: 21px;
-  height: 21px;
+  width: 22px;
+  height: 22px;
   border-radius: 50%;
   background-color: var(--sub-color1);
 }
@@ -80,12 +108,13 @@
   position: relative;
 }
 
-.selected abbr {
+.selected:not(.react-calendar__month-view__days__day--neighboringMonth) abbr {
   position: relative;
   z-index: 10;
   color: white;
 }
 
+/* 연속 선택된 날짜 스타일 */
 .selected-continuous {
   background-color: var(--sub-color1);
   color: white;
@@ -105,12 +134,8 @@
   border-bottom-right-radius: 50px;
 }
 
+/* 이전/이후 달 날짜 스타일 */
 .react-calendar__month-view__days__day--neighboringMonth {
-  background-color: white;
+  background: white;
   color: #757575;
-}
-
-.react-calendar__tile--now:enabled:hover {
-  background: var(--sub-color1);
-  color: white;
 }
